+++ taxel-gui/CHANGELOG.md
@@ -7,4 +7,5 @@
   - Display facts in main table
   - Display sections in sidebar
   - Select display level
+  - Select label language
   - Clear main table